requirements.txt
django>=3.1.6
djangorestframework>=3.12.2
drf-access-policy>=0.8.7
redis
django-model-utils>=4.1.1
drf-yasg[validation]==1.20.0
django_json_widget>=1.1.1
